feat: unify loading components (#748)
