+++ packages/react-kit/src/components/cta/dispute/RemoveFeesDisputeResolverButton.tsx
@@ -5,7 +5,7 @@ import { Button } from "../../buttons/Button";
 import { useCoreSdk } from "../../../hooks/core-sdk/useCoreSdk";
 import { ButtonTextWrapper, ExtraInfo, LoadingWrapper } from "../common/styles";
 import { CtaButtonProps } from "../common/types";
-import { Loading } from "../../Loading";
+import { Loading } from "../../ui/loading/Loading";
 import { CreateSellerArgs, TransactionResponse } from "@bosonprotocol/common";
 import { ButtonSize } from "../../ui/buttonSize";
 import { withQueryClientProvider } from "../../queryClient/withQueryClientProvider";
