+++ packages/react-kit/src/components/modal/components/Transactions/WaitingForConfirmationModal/WaitingForConfirmationModal.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import { theme } from "../../../../../theme";
 
 import { Grid } from "../../../../ui/Grid";
-import Loading from "../../../../ui/loading/Loading";
+import Loading from "../../../../ui/loading/LoadingWrapper";
 import { Typography } from "../../../../ui/Typography";
 import { GenericModalProps } from "../../../ModalContext";
 import { useModal } from "../../../useModal";
