+++ packages/react-kit/src/index.tsx
@@ -32,7 +32,7 @@ export * from "./components/cta/funds/WithdrawAllFundsButton";
 export * from "./components/currencyDisplay/CurrencyDisplay";
 export * from "./components/exchangeCard/ExchangeCard";
 export * from "./components/image/Image";
-export * from "./components/Loading";
+export * from "./components/ui/loading/Loading";
 export * from "./components/productCard/ProductCard";
 export * from "./components/productCard/const";
 export * from "./components/skeleton/ProductCardSkeleton";
